requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
dataclasses-json==0.5.1
fake-useragent==0.1.11
Flask==1.1.2
Flask-Cors==3.0.8
future==0.18.2
get==2019.4.13
idna==2.9
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.1
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-enum==1.5.1
MechanicalSoup==0.12.0
mypy-extensions==0.4.3
oauthlib==3.1.0
parse==1.15.0
post==2019.4.13
praw==7.1.0
prawcore==1.4.0
public==2019.4.13
pyee==7.0.2
pyppeteer==0.2.2
pyquery==1.4.1
python-youtube==0.6.0
pytz==2020.1
query-string==2019.4.13
request==2019.4.13
requests==2.24.0
requests-html==0.10.0
requests-oauthlib==1.3.0
responses==0.10.15
six==1.15.0
soupsieve==2.0.1
stringcase==1.2.0
tqdm==4.46.1
twitter-scraper==0.4.2
typing-extensions==3.7.4.2
typing-inspect==0.6.0
update-checker==0.17
urllib3==1.25.9
w3lib==1.22.0
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
python-dateutil==2.8.1
gunicorn==19.0.0
